feat: ensure Mill is installed
**The problem**

So when using this action I was a little thrown off that it was failing
with:

```
java.io.IOException: Cannot run program "mill" (in directory "/home/runner/scala-steward/workspace/repos/scalacenter/bloop-config"): error=2, No such file or directory
```

I would expect the action to ensure what's needed is installed. I'm
assuming no one hits on this for sbt due to it always being there in the
ubuntu image.

**The fix**

This change ensures that `mill` is always there and installed.