Display GitHub Action Job Summary
This is the complementary PR for https://github.com/scala-steward-org/scala-steward/pull/3071 -
while that PR _creates_ the summary report, this is the PR that links
up to GitHub's `core.summary` workflow command, to _display_ the report
in GitHub's Job summary.

https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#:~:text=core.summary
https://github.blog/2022-05-09-supercharging-github-actions-with-job-summaries/#:~:text=core.summary

In this change, if the `run-summary.md` file generated by the
Scala Steward JVM process is found, it's written to the GitHub  summary.