Don't persist the summary beyond its own run
This was something we missed from
https://github.com/scala-steward-org/scala-steward-action/pull/502 -
because the `run-summary.md` file is saved by
https://github.com/scala-steward-org/scala-steward/pull/3071
to the Scala Steward workspace, and almost the entire workspace is
persisted to cache by Scala Steward's GitHub Action, the report would
be persisted until the _next_ run of the GitHub Action, and
concievably, if the next run went wrong somehow, the incorrect _old_
summary could get reported against that subsequent run, which would
be bad.

The fix here is to delete the summary at the `saveWorkspaceCache()`
step, which occurs after the summary has been transferred to the
`GITHUB_STEP_SUMMARY` file by calling `core.summary()`.